Execute Python script in file dir (as cwd)
Previously, cwd will always be the root of the repo and so absolute pathing to the input files needed to be specified. Now VSCode is set to execute the Python script in the folder it is located in as cwd and so only relative pathing to the input file needs to be specified.